add new token for each request
